move ureg references to units.py
