+++ opgee/processes/crude_oil_dewatering.py
@@ -6,12 +6,12 @@
 # Copyright (c) 2021-2022 The Board of Trustees of the Leland Stanford Junior University.
 # See LICENSE.txt for license details.
 #
-from .. import ureg
 from ..core import TemperaturePressure
 from ..error import OpgeeException
 from ..log import getLogger
 from ..process import Process
 from ..stream import PHASE_LIQUID
+from ..units import ureg
 from .shared import get_energy_carrier
 
 _logger = getLogger(__name__)
