+++ opgee/processes/demethanizer.py
@@ -8,16 +8,15 @@
 #
 import pandas as pd
 
-from .. import ureg
 from ..core import STP, TemperaturePressure
 from ..emissions import EM_FUGITIVES
 from ..energy import EN_ELECTRICITY
 from ..log import getLogger
-from ..process import Process
-from ..process import run_corr_eqns
+from ..process import Process, run_corr_eqns
 from ..stream import PHASE_GAS, Stream
+from ..units import ureg
 from .compressor import Compressor
-from .shared import get_energy_carrier, predict_blower_energy_use, get_bounded_value
+from .shared import get_bounded_value, get_energy_carrier, predict_blower_energy_use
 
 _logger = getLogger(__name__)
