+++ opgee/processes/ryan_holmes.py
@@ -6,12 +6,12 @@
 # Copyright (c) 2021-2022 The Board of Trustees of the Leland Stanford Junior University.
 # See LICENSE.txt for license details.
 #
-from .. import ureg
 from ..emissions import EM_FUGITIVES
-from ..energy import EN_NATURAL_GAS, EN_DIESEL
+from ..energy import EN_DIESEL, EN_NATURAL_GAS
 from ..log import getLogger
 from ..process import Process
 from ..stream import PHASE_GAS
+from ..units import ureg
 
 _logger = getLogger(__name__)
