+++ opgee/processes/sour_gas_compressor.py
@@ -6,11 +6,11 @@
 # Copyright (c) 2021-2022 The Board of Trustees of the Leland Stanford Junior University.
 # See LICENSE.txt for license details.
 #
-from .. import ureg
 from ..emissions import EM_FUGITIVES
 from ..log import getLogger
 from ..process import Process
 from ..processes.compressor import Compressor
+from ..units import ureg
 from .shared import get_energy_carrier
 
 _logger = getLogger(__name__)
