+++ opgee/processes/venting.py
@@ -6,11 +6,11 @@
 # Copyright (c) 2021-2022 The Board of Trustees of the Leland Stanford Junior University.
 # See LICENSE.txt for license details.
 #
-from .. import ureg
-from ..emissions import EM_VENTING, EM_FUGITIVES
+from ..emissions import EM_FUGITIVES, EM_VENTING
 from ..log import getLogger
 from ..process import Process
 from ..stream import Stream
+from ..units import ureg
 
 _logger = getLogger(__name__)
