+++ opgee/processes/water_treatment.py
@@ -6,13 +6,13 @@
 # Copyright (c) 2021-2022 The Board of Trustees of the Leland Stanford Junior University.
 # See LICENSE.txt for license details.
 #
-from .. import ureg
 from ..core import TemperaturePressure
 from ..energy import EN_ELECTRICITY
 from ..error import OpgeeException
 from ..import_export import WATER
 from ..log import getLogger
 from ..process import Process
+from ..units import ureg
 
 _logger = getLogger(__name__)
