+++ tests/test_emissions.py
@@ -1,6 +1,6 @@
 import pytest
 import pandas as pd
-from opgee import ureg
+from opgee.units import ureg
 from opgee.emissions import Emissions, EM_FUGITIVES, EM_FLARING, EM_LAND_USE, EmissionsError
 from opgee.error import OpgeeException
