+++ tests/test_energy.py
@@ -1,4 +1,4 @@
-from opgee import ureg
+from opgee.units import ureg
 from opgee.energy import (Energy, EN_DIESEL, EN_NATURAL_GAS, EN_RESID,
                           EN_PETCOKE, EN_CRUDE_OIL, EN_ELECTRICITY)
 from opgee.error import OpgeeException
