+++ tests/test_import_export.py
@@ -1,4 +1,4 @@
-from opgee import ureg
+from opgee.units import ureg
 from opgee.import_export import ImportExport, ELECTRICITY, NATURAL_GAS
 
 def test_import():
