+++ tests/test_impute.py
@@ -2,7 +2,7 @@ import pytest
 from opgee.error import OpgeeException
 from opgee.process import Process
 from .utils_for_tests import load_test_model
-from opgee import ureg
+from opgee.units import ureg
 
 class CopyingProcess(Process):
     def run(self, analysis):
