+++ tests/test_opgee_xml.py
@@ -1,6 +1,6 @@
 import pytest
 
-from opgee import ureg
+from opgee.units import ureg
 from opgee.model_file import ModelFile
 from tests.utils_for_tests import path_to_test_file
