+++ tests/test_process_loop.py
@@ -1,4 +1,4 @@
-from opgee import ureg
+from opgee.units import ureg
 from opgee.stream import Stream, PHASE_LIQUID
 from opgee.process import Process
 from .utils_for_tests import load_test_model
