+++ tests/test_smart_defaults.py
@@ -1,4 +1,4 @@
-from opgee import ureg
+from opgee.units import ureg
 from .utils_for_tests import load_model_from_str
 
 template = """<?xml version='1.0' encoding='UTF-8'?>
