+++ tests/test_thermofunction.py
@@ -2,7 +2,7 @@ import pandas as pd
 import pytest
 from opgee.core import TemperaturePressure
 from opgee.stream import Stream, PHASE_GAS, PHASE_LIQUID
-from opgee import ureg
+from opgee.units import ureg
 
 
 @pytest.fixture(scope="module")
